VIL-979: move checkSSO into @server/lib/check-sso
