+++ src/server-actions/authentication/login.ts
@@ -2,11 +2,10 @@
 
 import { type User } from '@server/database/schemas/users';
 import { auth } from '@server/lib/auth';
+import { checkSSO } from '@server/lib/check-sso';
 import { getStringValue } from '@server/lib/get-string-value';
 import { redirect, RedirectType } from 'next/navigation';
 import { getExtracted } from 'next-intl/server';
-
-import { checkSSO } from './check-sso';
 
 export async function login(_previousState: string, formData: FormData): Promise<string> {
     const t = await getExtracted('common');
